Add Maxweight to repository
Given an array of integers.  All but one integer is the same number.
(it's bigger.)  Find the index that has the bigger number.  (Note:  You
are not use the max() function, but you can assume that there is only
ONE number that is bigger than the other.)

This problem is solved thru a binary search.  The project is tested thru
TestNG and built thru Maven.